requirements.txt
Django>=2.2.7,<2.3.0
boto3==1.10.28
gunicorn==19.9.0
whitenoise==4.1.2
djangorestframework>=3.10.3,<3.11.0
django-storages==1.8
django-cors-headers>=3.2.0,<3.3.0
psycopg2>=2.8.4,<2.9.0
django-heroku==0.3.1
Pillow>=6.2.1,<6.3.0